// packages/cli/src/acp-integration/schema.d.ts
export type ModelInfo = z.infer<typeof modelInfoSchema>;
// ...
export declare const modelInfoSchema: z.ZodObject<{
    _meta: z.ZodOptional<z.ZodNullable<z.ZodRecord<z.ZodString, z.ZodUnknown>>>;
    description: z.ZodOptional<z.ZodNullable<z.ZodString>>;
    modelId: z.ZodString;
    name: z.ZodString;
}, "strip", z.ZodTypeAny, {
    name: string;
    modelId: string;
    description?: string | null | undefined;
    _meta?: Record<string, unknown> | null | undefined;
}, {
    name: string;
    modelId: string;
    description?: string | null | undefined;
    _meta?: Record<string, unknown> | null | undefined;
}>;